Fixed handling of the 'dash' special character.
Dragon 14 seems to send an 'en dash' when using 'dash'. We already handle it as a standalone character, but when used as part of a dictation sequence, Dragon controls the character. By the time it reaches us, it's in extended ASCII as 0x96, which breaks dragonfly as dragonfly is expecting UTF-8 strings. The simple fix is to replace the 'en dash' with a hyphen. While not technically the same, this entire project is set up for programmers and it's more than likely the intended behavior.